# api/openapi.yaml
openapi: 3.1.0
info:
  title: SYNC Server API
  version: 0.1.0
  description: |
    Federated, privacy-focused 1:1 chat server.
    Access tokens are short-lived HS256 JWTs (1 h TTL).
    Refresh tokens are 30-day opaque tokens subject to rotation and family-replay detection.
servers:
  - url: http://localhost:8080
    description: Local development
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    GuildSnapshot:
      type: object
      required:
        - active_days
        - level
        - contribution_score
        - rank
        - level_progress_percent
        - daily_outbound_messages_enforced
        - daily_outbound_messages_sent
        - daily_attachment_sends_enforced
        - daily_attachment_sends_sent
        - allowed_attachment_types
        - daily_friend_adds_enforced
        - daily_friend_adds_sent
        - challenge_state
      properties:
        active_days:
          type: integer
        level:
          type: integer
          minimum: 1
          maximum: 10
        contribution_score:
          type: integer
        rank:
          type: string
          enum: [F, E, D, C, B, A, S]
        next_level_active_days:
          type: integer
          nullable: true
        level_progress_percent:
          type: integer
          minimum: 0
          maximum: 100
        daily_outbound_messages_enforced:
          type: boolean
        daily_outbound_messages_limit:
          type: integer
          nullable: true
        daily_outbound_messages_sent:
          type: integer
        daily_outbound_messages_remaining:
          type: integer
          nullable: true
        daily_attachment_sends_enforced:
          type: boolean
        daily_attachment_send_limit:
          type: integer
          nullable: true
        daily_attachment_sends_sent:
          type: integer
        daily_attachment_sends_remaining:
          type: integer
          nullable: true
        allowed_attachment_types:
          type: array
          items:
            type: string
        daily_friend_adds_enforced:
          type: boolean
        daily_friend_add_limit:
          type: integer
          nullable: true
        daily_friend_adds_sent:
          type: integer
        daily_friend_adds_remaining:
          type: integer
          nullable: true
        challenge_state:
          type: string
          enum: [none, challenged, frozen]
          description: >
            Client-visible progression challenge state.
            "none" – no challenge active.
            "challenged" – suspicious activity detected; progression paused pending clean activity.
            "frozen" – repeated suspicious attempts; progression frozen until review or quiet-period clears.
        pending_milestone_notification:
          allOf:
            - $ref: '#/components/schemas/GuildMilestoneNotification'
          nullable: true
          description: >
            Non-null when a level-up, rank-up, or attachment-type unlock was detected on this
            request.  The client is responsible for dismissing the notification locally; the server
            does not persist per-user notification state.  Safe to ignore.
    GuildMilestoneNotification:
      type: object
      description: >
        Payload the client uses to render a milestone toast, badge, or unlock animation.
        Returned as `pending_milestone_notification` in `GuildSnapshot` when a server-side
        progression event was detected on this request.  Null when no milestone is pending.
        The client is responsible for dismissing / persisting the notification locally;
        the server does not maintain per-user notification state.
      required:
        - kind
        - badge_label
        - headline_key
        - detail_key
        - new_value
      properties:
        kind:
          type: string
          enum: [level_up, rank_up, unlock_attachment_type]
          description: >
            Kind of progression milestone that triggered this notification.
            "level_up" – the user reached a new guild level on this request.
            "rank_up" – the user's contribution score crossed a rank threshold.
            "unlock_attachment_type" – a new attachment MIME type became available.
        badge_label:
          type: string
          description: Human-readable badge label, e.g. "Level 5" or "Rank D".
        headline_key:
          type: string
          description: Localisation key for the unlock headline, e.g. "guild.milestone.level_up".
        detail_key:
          type: string
          description: Localisation key for the unlock sub-copy, e.g. "guild.milestone.level_5_detail".
        unlocked_value:
          type: string
          nullable: true
          description: >
            Optional specific thing unlocked, e.g. "application/pdf" for an attachment-type unlock.
            Null for level-up and rank-up milestones.
        new_value:
          type: string
          description: >
            New level or rank value after the milestone, e.g. "5" for a level-up or "D" for a rank-up.
...
    UserProfilePublic:
      type: object
      required: [id, username]
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        avatar_base64:
          type: string
          nullable: true
        message_public_key:
          type: string
          nullable: true
        guild:
          allOf:
            - $ref: '#/components/schemas/GuildSnapshot'
          nullable: true
          description: Returned for `/api/profile/me`; omitted for public profile lookups.
    UpdateProfileRequest:
      type: object
      properties:
        username:
          type: string
          nullable: true
        avatar_base64:
          type: string
          nullable: true
          description: Base64-encoded avatar image. Pass `null` to clear the avatar.
        message_public_key:
          type: string
          nullable: true
          description: Base64-encoded 32-byte public key for message encryption.
...
    ErrorBody:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
  responses:
    Unauthorized:
      description: Missing or invalid JWT
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
...
    UnprocessableEntity:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
...
paths:
...
  /api/profile/me:
    get:
      operationId: getMyProfile
      summary: Get the current user's profile
      tags: [profile]
      responses:
        '200':
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfilePublic'
        '401':
          $ref: '#/components/responses/Unauthorized'
    patch:
      operationId: updateMyProfile
      summary: Update the current user's profile
      tags: [profile]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
      responses:
        '200':
          description: Updated user profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfilePublic'
        '400':
          $ref: '#/components/responses/UnprocessableEntity'
        '401':
          $ref: '#/components/responses/Unauthorized'
    delete:
      operationId: deleteMyProfile
      summary: Delete the current user's account
      tags: [profile]
      responses:
        '204':
          description: Account deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/profile/{user_id}:
    get:
      operationId: getUserProfile
      summary: Get a public user profile by ID
      tags: [profile]
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Public user profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfilePublic'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...